Move network related code into separate module
The network code that handles raw IP traffic and arp is now in a
separate net mod. The future plan is to pull this out into a separate
crate.
