+++ src/net/mod.rs
@@ -1,0 +1,7 @@
+//! Networking primitives
+//!
+
+pub mod arp;
+pub mod ether;
+pub mod ip;
+pub mod mac;
